board:2093 modernize-avoid-c-style-cast reaches ZERO, and the gate was not analysing everything
  this check      3667 -> 0
  findings       12979 -> 9300
  Venice      d99e6aaf   unchanged
  OldTown     ecb4e513   unchanged

THE HEADERS TOOK THREE ATTEMPTS AND THE THIRD TAUGHT SOMETHING ABOUT THE GATE.

run-clang-tidy with -fix collects every translation unit's replacements and writes them TOGETHER at
the end. A header read by twenty units gets its replacement applied twenty times, and the result is
static_cast<size_t>(term)) with one paren too many -- 140 broken lines across 39 files. Reverted.
clang-apply-replacements over -export-fixes does the same thing; reverted again.

What works is one translation unit at a time, written immediately: the next unit reads a header that
is already repaired and emits nothing for it.

But sequentially it still fixed almost nothing, and the reason is the finding: clang-tidy refuses to
apply ANY fix to a unit that carries an error, and most units carried

    ParticipatingMedium.h:4:10: error: 'numbers' file not found

The compile database's compiler is Apple's c++, which supplies its own sysroot implicitly. LLVM's
clang-tidy does not, so it cannot find libc++ at all. With -isysroot handed to it the same sweep
fixed 52 headers in one pass.

**The gate runs the same clang-tidy.** What it counts is what parses, and some of the tree may not
be parsing at all. That is the first trap CLAUDE.md names -- a gate blind to a path -- and it needs
measuring before any count from this gate is quoted as a total. Filed as the next thing to check.

The last seven were judgements a tool should not make: pthread_get_stackaddr_np and
__builtin_frame_address are pointers becoming integers, so reinterpret_cast; a painted stack word is
an integer becoming a pointer, likewise; a size becoming uint32_t is static_cast; and a buffer
handle tested for truth is static_cast<bool>.

+++ src/render/SceneRenderer.h
@@ -101,7 +101,7 @@ public:
   void CastsBelow(uint32_t slot) { Shadow_.CastsBelow(slot); }
 
   [[nodiscard]] ReadState ReadShadowAtlas(std::vector<float> &depth);
-  static constexpr float kNearM = (float)outshine::Camera::kNearestM;
+  static constexpr float kNearM = static_cast<float>(outshine::Camera::kNearestM);
 
   [[nodiscard]] ReadState ReadSceneLinear(std::vector<float> &rgba);
 
@@ -239,7 +239,9 @@ public:
     uint32_t Layouts = 0;
   };
 
-  [[nodiscard]] const Effort &Spent(Stage stage) const { return Spent_[(size_t)stage]; }
+  [[nodiscard]] const Effort &Spent(Stage stage) const {
+    return Spent_[static_cast<size_t>(stage)];
+  }
 
   [[nodiscard]] size_t SubjectUniformPushes() const {
     return Subjects_.UniformPushes() + Glass_.UniformPushes();
@@ -256,7 +258,7 @@ public:
                       const double right[3],
                       const double up[3]);
 
-  void SetFovDeg(double deg) { FovDeg_ = deg > 0.0 ? (float)deg : FovDeg_; }
+  void SetFovDeg(double deg) { FovDeg_ = deg > 0.0 ? static_cast<float>(deg) : FovDeg_; }
 
   [[nodiscard]] bool SetGroundClasses(const uint32_t *words,
                                       size_t wordCount,
@@ -264,9 +266,9 @@ public:
                                       size_t paletteFloats,
                                       std::string &error);
 
-  void SetOrthoM(double m) { OrthoM_ = (float)m; }
+  void SetOrthoM(double m) { OrthoM_ = static_cast<float>(m); }
 
-  void SetNearM(double m) { NearM_ = m > 0.0 ? (float)m : NearM_; }
+  void SetNearM(double m) { NearM_ = m > 0.0 ? static_cast<float>(m) : NearM_; }
 
   [[nodiscard]] float NearMetres() const { return NearM_; }
 
@@ -280,7 +282,7 @@ public:
   [[nodiscard]] double PictureH() const;
 
   [[nodiscard]] double SceneAspect() const {
-    return PictureH() > 0 ? (double)PictureW() / (double)PictureH() : 0.0;
+    return PictureH() > 0 ? PictureW() / PictureH() : 0.0;
   }
 
 private:
